BDW/pyproject.toml
[tool.poetry]
name = "BDW"
version = "0.1.0"
description = "A bad discord wrapper for python."
authors = ["Fox551 <BadPythonCoder@TheInternet.net>","TBSharedAccount <TBSharedAccount@TheInternet.net"]

[tool.poetry.dependencies]
python = "^3.8"
websocket-client = "1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
